@@ -3,16 +3,20 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
+        "@chakra-ui/icons": "^2.0.2",
+        "@chakra-ui/react": "^2.2.1",
+        "@emotion/react": "^11.9.3",
+        "@emotion/styled": "^11.9.3",
         "@testing-library/jest-dom": "^5.16.2",
         "@testing-library/react": "^12.1.3",
         "@testing-library/user-event": "^13.5.0",
         "agora-react-uikit": "^1.0.7",
         "axios": "^0.26.1",
-        "framer-motion": "^6.3.3",
+        "framer-motion": "^6.3.13",
         "lottie-react": "^2.2.1",
         "otp-input-react": "^0.3.0",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "react-loading": "^2.0.3",
         "react-phone-number-input": "^3.1.53",
         "react-router-dom": "5.1.0",
