@@ -1,13 +1,15 @@
+import { ChakraProvider } from '@chakra-ui/react';
 import * as React from 'react'
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client'
 import App from './App';
 import { PhoneNumberContextProvider } from './context/phoneNumberContext';
 
-ReactDOM.render(
+createRoot(document.getElementById('root')).render(
   <React.StrictMode>
+    <ChakraProvider>
       <PhoneNumberContextProvider>
         <App />
       </PhoneNumberContextProvider>
+    </ChakraProvider>
   </React.StrictMode>,
-  document.getElementById('root')
 );
