@@ -3,152 +3,136 @@ import ProfileModal from '../ProfileModal/ProfileModal';
 import "./navbar.scss"
 import { useHistory } from "react-router-dom";
 import { PhoneNumberContext } from '../../context/phoneNumberContext';
-import {  motion } from 'framer-motion';
+import { motion } from 'framer-motion';
+import {
+  Box,
+  Flex,
+  Avatar,
+  Link,
+  Button,
+  Menu,
+  MenuButton,
+  MenuList,
+  MenuItem,
+  MenuDivider,
+  useDisclosure,
+  useColorModeValue,
+  Stack,
+  useColorMode,
+  Center,
+  Portal
+} from '@chakra-ui/react';
+import { BellIcon, MoonIcon, SunIcon } from '@chakra-ui/icons';
+
+const NavLink = ({ children }) => (
+  <Link
+    px={2}
+    py={1}
+    rounded={'md'}
+    _hover={{
+      textDecoration: 'none',
+      bg: useColorModeValue('gray.200', 'gray.700'),
+    }}
+    href={'#'}>
+    {children}
+  </Link>
+);
 
 const Navbar = () => {
-
-  const [transformmm, setTransformmm] = React.useState(false);
+  const { colorMode, toggleColorMode } = useColorMode();
 
   const user = JSON.parse(localStorage.getItem('user'));
 
-  const { notification, dispatch, mobile } = React.useContext(PhoneNumberContext);
+  const { notification, dispatch } = React.useContext(PhoneNumberContext);
   // console.log(notification);
 
-  const [show, setShow] = React.useState(false);
-
   let history = useHistory();
-
-  const [profile, setProfile] = React.useState(false);
-
-  useEffect(() => {
-    setTimeout(() => {
-      if (show) {
-        setShow(!show);
-      }
-    }, 10000);
-  }, [show])
-
 
   const handleLogout = () => {
     localStorage.removeItem('user');
     history.push('/');
   }
 
-  const handleProfile = () => {
-    setProfile(!profile);
-  }
-
-  const handleNotification = () => {
-    setShow(!show);
-  }
-
-  const list = {
-    visible: {
-      opacity: 1,
-      transition: {
-        when: "beforeChildren",
-        staggerChildren: 0.3,
-      },
-    },
-    hidden: {
-      opacity: 0,
-      transition: {
-        when: "afterChildren",
-      },
-    },
-  }
-
-  const item = {
-    visible: { opacity: 1, x: 0 },
-    hidden: { opacity: 0, x: -100 },
-  }
-
   return (
-    <div className='navbar'>
-      {mobile ?
-        <img
-          src="https://img.icons8.com/material-outlined/24/000000/menu--v1.png" alt='menu1' className='menu' onClick={() => dispatch({ type: 'SET_MOBILE' })} />
-        :
-        <img
-          src="https://img.icons8.com/ios/50/000000/delete-sign--v1.png" alt='menu1' className='menu' onClick={() => dispatch({ type: 'SET_MOBILE' })} />
-      }
-      <div className="profile">
-        <img src={user.pic} alt="avatar" className='avatar' onClick={handleProfile} />
-        <p>{user.username}</p>
-        <img src="https://img.icons8.com/ios/50/000000/appointment-reminders--v1.png" alt='notification' style={{
-          transform: transformmm ? 'scale(1.1) translateY(-2px)' : null,
-          transition: 'all 0.1s ease-in-out',
-          cursor: 'pointer',
-          marginRight: '30px',
-          height: '40px',
-        }}
-          onClick={handleNotification}
-          onMouseEnter={() => setTransformmm(true)}
-          onMouseLeave={() => setTransformmm(false)}
-        />
-        {
-          notification?.length > 0 &&
-          <div className="number" >
-            <p>{notification?.length}</p>
-          </div>
-        }
-      </div>
-      {show ? <div className="notificationModal">
-        <motion.ul
-          initial="hidden"
-          animate="visible"
-          variants={list}
-        >
-          {!notification.length ? <li onClick={handleNotification}>No new notifications</li> : notification.map((notifications) => {
-            return (
-              <>
-                <li
-                  variants={item}
-                  key={notifications._id}
-                  onClick={() => {
-                    console.log(notifications);
-                    setShow(!show);
-                    dispatch({
-                      type: 'SET_SELECTED_CHAT',
-                      payload: notifications.chat
-                    })
-                    dispatch({
-                      type: 'SET_NOTIFICATION',
-                      payload: notifications.filter(notifications._id !== notification._id)
-                    })
+    <> 
+      <Box px={4}>
+        <Flex h={16} alignItems={'center'} justifyContent={'space-between'}>
+          <Box>ChatApp</Box>
 
-                    //check this feature
-                  }}>
-                  {notifications.chat.isGroupChat ? `New Message in ${notifications.chat.chatName}` : `New Message from ${notifications.sender.username}`}
-                </li>
-                <hr />
-              </>
-            )
-          })}
-        </motion.ul>
-      </div> : null}
-      {profile ?
-        <div className="myProfile">
-          <motion.ul
-            initial="hidden"
-            animate="visible"
-            variants={list}
-          >
-            <ProfileModal user={user} p={setProfile}>
-              <li
-                variants={item}>
-                My Profile
-              </li>
-            </ProfileModal>
-            <hr />
-            <li
-              variants={item}
-              onClick={handleLogout}
-            >Logout</li>
-          </motion.ul>
-        </div>
-        : null}
-    </div>
+          <Flex alignItems={'center'}>
+            <Stack direction={'row'} spacing={7}>
+              <Button bg={colorMode === 'light' ? 'white' : 'blackAlpha.600'} onClick={toggleColorMode}>
+                {colorMode === 'light' ? <MoonIcon /> : <SunIcon />}
+              </Button>
+              <Menu>
+                <MenuButton>
+                  <BellIcon fontSize={'xl'} />
+                </MenuButton>
+                <Portal>
+                  <MenuList>
+                    {!notification.length
+                      ? <MenuItem>No new notifications</MenuItem>
+                      :
+                      notification.map((notifications) =>
+                        <MenuItem
+                          key={notifications._id}
+                          onClick={() => {
+                            console.log(notifications);
+                            dispatch({
+                              type: 'SET_SELECTED_CHAT',
+                              payload: notifications.chat
+                            })
+                            dispatch({
+                              type: 'SET_NOTIFICATION',
+                              payload: notifications.filter(notifications._id !== notification._id)
+                            })
+                            //check this feature
+                          }
+                          }>
+                          {notifications.chat.isGroupChat ? `New Message in ${notifications.chat.chatName}` : `New Message from ${notifications.sender.username}`}
+                        </MenuItem>
+                      )}
+                  </MenuList>
+                </Portal>
+              </Menu>
+
+              <Menu>
+                <MenuButton
+                  as={Button}
+                  rounded={'full'}
+                  variant={'link'}
+                  cursor={'pointer'}
+                  minW={0}>
+                  <Avatar
+                    size={'sm'}
+                    src={user.pic}
+                  />
+                </MenuButton>
+                <MenuList alignItems={'center'}>
+                  <br />
+                  <Center>
+                    <Avatar
+                      size={'2xl'}
+                      src={user.pic}
+                    />
+                  </Center>
+                  <br />
+                  <Center>
+                    <p>{user.username}</p>
+                  </Center>
+                  <br />
+                  <MenuDivider />
+                  <ProfileModal user={user}>
+                    <MenuItem>My Profile</MenuItem>
+                  </ProfileModal>
+                  <MenuItem onClick={handleLogout}>Logout</MenuItem>
+                </MenuList>
+              </Menu>
+            </Stack>
+          </Flex>
+        </Flex>
+      </Box>
+    </>
   )
 }
 
